@@ -31,58 +30,47 @@
     "postinstall": "ngcc --properties es2015 es5 browser module main --first-only --create-ivy-entry-points --tsconfig \"./src/tsconfig.app.json\""
   },
   "dependencies": {
-    "@akveo/ng2-completer": "^9.0.1",
-    "@angular/animations": "^13.3.12",
-    "@angular/cdk": "12.1.0",
-    "@angular/common": "^13.3.12",
-    "@angular/compiler": "^13.3.12",
-    "@angular/core": "^13.3.12",
-    "@angular/forms": "^13.3.12",
+    "@angular/animations": "^15.2.10",
+    "@angular/cdk": "^15.2.9",
+    "@angular/common": "^15.2.10",
+    "@angular/compiler": "^15.2.10",
+    "@angular/core": "^15.2.10",
+    "@angular/forms": "^15.2.10",
     "@angular/google-maps": "^12.2.13",
-    "@angular/platform-browser": "^13.3.12",
-    "@angular/platform-browser-dynamic": "^13.3.12",
-    "@angular/router": "^13.3.12",
-    "@nebular/auth": "9.1.0-rc.8",
-    "@nebular/eva-icons": "9.1.0-rc.8",
-    "@nebular/security": "9.1.0-rc.8",
-    "@nebular/theme": "9.1.0-rc.8",
-    "@ngx-translate/core": "^14.0.0",
-    "@ngx-translate/http-loader": "^7.0.0",
+    "@angular/platform-browser": "^15.2.10",
+    "@angular/platform-browser-dynamic": "^15.2.10",
+    "@angular/router": "^15.2.10",
+    "@asymmetrik/ngx-leaflet": "3.0.1",
+    "@nebular/auth": "11.0.1",
+    "@nebular/eva-icons": "11.0.1",
+    "@nebular/security": "11.0.1",
+    "@nebular/theme": "11.0.1",
     "@swimlane/ngx-charts": "^14.0.0",
     "@terraformer/spatial": "^2.1.2",
     "@terraformer/wkt": "^2.1.2",
-    "@types/leaflet-draw": "^1.0.6",
-    "angular-notifier": "^11.0.0",
     "angular2-chartjs": "0.4.1",
-    "asn1js": "^3.0.5",
     "bootstrap": "4.3.1",
-    "chart.js": "2.9.4",
+    "chart.js": "2.7.1",
     "ckeditor": "4.7.3",
     "classlist.js": "1.1.20150312",
     "core-js": "2.5.1",
-    "dot": "^1.1.3",
-    "echarts": "^5.4.2",
+    "echarts": "^4.9.0",
     "eva-icons": "^1.1.3",
     "file-saver": "^2.0.5",
     "intl": "1.2.5",
     "ionicons": "2.0.1",
-    "json-schema": "^0.4.0",
-    "keycloak-angular": "9.3.2",
-    "keycloak-js": "17.0.1",
-    "leaflet": "^1.9.2",
+    "keycloak-angular": "13.1.0",
+    "keycloak-js": "18.0.1",
+    "leaflet": "1.2.0",
     "leaflet-draw": "^1.0.4",
-    "lodash": "^4.17.21",
-    "lodash.template": "^4.5.0",
     "lucene-query-string-builder": "^1.0.8",
-    "minimist": "^1.2.8",
     "nebular-icons": "1.1.0",
     "ng2-ckeditor": "~1.2.9",
     "ng2-completer": "^9.0.1",
     "ng2-smart-table": "^1.6.0",
-    "ngx-autosize-input": "^13.1.1",
+    "ngx-autosize-input": "^15.0.4",
     "ngx-echarts": "^4.2.2",
-    "ngx-flag-picker": "^12.0.0",
-    "node-sass": "^8.0.0",
+    "node-sass": "^4.14.1",
     "normalize.css": "6.0.0",
     "pace-js": "1.0.2",
     "pkijs": "^3.0.14",
@@ -102,28 +90,32 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^16.2.8",
-    "@angular-eslint/builder": "13.5.0",
-    "@angular-eslint/eslint-plugin": "13.5.0",
-    "@angular-eslint/eslint-plugin-template": "13.5.0",
-    "@angular-eslint/schematics": "13.5.0",
-    "@angular-eslint/template-parser": "13.5.0",
-    "@angular/cli": "^13.3.10",
-    "@angular/compiler-cli": "^13.3.12",
-    "@angular/language-service": "13.3.12",
-    "@compodoc/compodoc": "1.1.19",
+    "@angular-devkit/build-angular": "^15.2.10",
+    "@angular-eslint/builder": "15.2.1",
+    "@angular-eslint/eslint-plugin": "15.2.1",
+    "@angular-eslint/eslint-plugin-template": "15.2.1",
+    "@angular-eslint/schematics": "15.2.1",
+    "@angular-eslint/template-parser": "15.2.1",
+    "@angular/cli": "^15.2.10",
+    "@angular/compiler-cli": "^15.2.10",
+    "@angular/language-service": "15.2.10",
+    "@compodoc/compodoc": "1.0.1",
     "@fortawesome/fontawesome-free": "^5.2.0",
+    "@ngx-translate/core": "^14.0.0",
+    "@ngx-translate/http-loader": "^7.0.0",
     "@types/d3-color": "1.0.5",
     "@types/jasmine": "~3.3.0",
     "@types/jasminewd2": "2.0.3",
-    "@types/leaflet": "^1.2.3",
+    "@types/leaflet": "1.2.3",
+    "@types/leaflet-draw": "^1.0.11",
     "@types/node": "^12.12.70",
-    "@typescript-eslint/eslint-plugin": "^5.54.1",
-    "@typescript-eslint/parser": "^5.54.1",
+    "@typescript-eslint/eslint-plugin": "^5.43.0",
+    "@typescript-eslint/parser": "^5.43.0",
     "angular-cli-ghpages": "^1.0.5",
+    "angular-notifier": "^13.0.0",
     "codelyzer": "^6.0.2",
-    "conventional-changelog-cli": "2.2.2",
-    "eslint": "^8.17.0",
+    "conventional-changelog-cli": "1.3.4",
+    "eslint": "^8.28.0",
     "husky": "0.13.3",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
@@ -133,15 +125,13 @@
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.2",
     "karma-jasmine-html-reporter": "^1.7.0",
-    "loader-utils": "^3.2.1",
     "npm-run-all": "4.0.2",
     "protractor": "~7.0.0",
     "rimraf": "2.6.1",
-    "socket.io-parser": "^4.2.3",
-    "stylelint": "15.11.0",
+    "stylelint": "7.13.0",
     "ts-node": "3.2.2",
     "tslint": "~6.1.0",
     "tslint-language-service": "^0.9.9",
-    "typescript": "~4.6.4"
+    "typescript": "~4.9.5"
   }
 }
