@@ -1,19 +1,3 @@
-/*
- * Copyright (c) 2022 Maritime Connectivity Platform Consortium
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 /**
  * @license
  * Copyright Akveo. All Rights Reserved.
@@ -23,7 +7,6 @@ import { Component, OnInit } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { AnalyticsService } from './@core/utils/analytics.service';
 import { SeoService } from './@core/utils/seo.service';
-import { addLangs } from './util/translateHelper';
 
 @Component({
   selector: 'ngx-app',
@@ -34,9 +17,7 @@ export class AppComponent implements OnInit {
   constructor(
     private analytics: AnalyticsService,
     private seoService: SeoService,
-    public translate: TranslateService,
-    ) {
-      addLangs(translate);
+    public translate: TranslateService,) {
   }
 
   ngOnInit(): void {
