@@ -1,35 +1,23 @@
-/*
- * Copyright (c) 2023 Maritime Connectivity Platform Consortium
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+/**
+ * @license
+ * Copyright Akveo. All Rights Reserved.
+ * Licensed under the MIT License. See License.txt in the project root for license information.
  */
-
- import { LandingModule } from './landing/landing.module';
- import { ApiModule as MIRApiModule } from './backend-api/identity-registry';
- import { ApiModule as MSRApiModule } from './backend-api/service-registry';
- import { ApiModule as SECOMApiModule } from './backend-api/secom';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { CoreModule } from './@core/core.module';
 import { ThemeModule } from './@theme/theme.module';
-import { AppComponent } from './app.component';
-import { customNotifierOptions } from './shared/customNotifierOption';
-import { NotifierModule } from 'angular-notifier';
-import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
+import { LandingModule } from './landing/landing.module';
+import { ApiModule as MIRApiModule } from './backend-api/identity-registry';
+import { ApiModule as MSRApiModule } from './backend-api/service-registry';
+import { ApiModule as SECOMApiModule } from './backend-api/secom';
+import { AuthModule } from './auth/auth.module';
+import { RouterModule } from '@angular/router';
 import {
   NbChatModule,
   NbDatepickerModule,
@@ -39,8 +27,9 @@ import {
   NbToastrModule,
   NbWindowModule,
 } from '@nebular/theme';
-import { AuthModule } from './auth/auth.module';
-import { RouterModule } from '@angular/router';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { NotifierModule } from 'angular-notifier';
+import { customNotifierOptions } from './shared/customNotifierOption';
 
 @NgModule({
   declarations: [AppComponent],
