@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Maritime Connectivity Platform Consortium
+ * Copyright (c) 2024 Maritime Connectivity Platform Consortium
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,6 +29,7 @@ export const initializeKeycloak = (keycloak: KeycloakService): () => Promise<boo
         initOptions: {
           onLoad: 'check-sso',
           flow: 'standard',
+          checkLoginIframe: false,
         },
         enableBearerInterceptor: true,
         bearerPrefix: 'Bearer',
