@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Maritime Connectivity Platform Consortium
+ * Copyright (c) 2024 Maritime Connectivity Platform Consortium
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -254,6 +254,7 @@ export const ColumnForResource = {
       title: 'Access type',
       type: 'string',
       description: 'OpenID Connect access type',
+      required: true,
       options: [
         {
           title: 'public',
@@ -683,8 +684,6 @@ export const ColumnForResource = {
       type: 'string',
       description: 'MCP MRN as unique identifer',
       visibleFrom: ['detail'],
-      shortIdType: 'instance',
-      immutable: true,
       required: true,
     },
     name: {
@@ -951,8 +950,6 @@ export const ColumnForResource = {
       type: 'string',
       description: 'MCP MRN as unique identifer',
       visibleFrom: ['detail'],
-      shortIdType: 'design',
-      immutable: true,
     },
     implementsServiceDesignVersion: {
       title: 'Technical design version',
