@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Maritime Connectivity Platform Consortium
+ * Copyright (c) 2024 Maritime Connectivity Platform Consortium
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -50,6 +50,11 @@ export const MIRResourceType: string[] = [ ResourceType.Device, ResourceType.Ves
   User = 'user',
   Vessel = 'vessel',
   MMS = 'mms',
+}
+
+export enum MrnAttributeInMSR {
+  Instance = 'instanceId',
+  Design = 'implementsServiceDesign',
 }
 
 /**
